Use %zu as the printf format for size_t.

+++ src/os_posix/os_map.c
@@ -33,12 +33,11 @@ __wt_mmap(WT_SESSION_IMPL *session,
 	    MAP_PRIVATE,
 	    fh->fd, (wt_off_t)0)) == MAP_FAILED) {
 		WT_RET_MSG(session, __wt_errno(),
-		    "%s map error: failed to map %" PRIuMAX " bytes",
-		    fh->name, (uintmax_t)orig_size);
+		    "%s map error: failed to map %zu bytes",
+		    fh->name, orig_size);
 	}
 	(void)__wt_verbose(session, WT_VERB_FILEOPS,
-	    "%s: map %p: %" PRIuMAX " bytes",
-	    fh->name, map, (uintmax_t)orig_size);
+	    "%s: map %p: %zu bytes", fh->name, map, orig_size);
 
 	*(void **)mapp = map;
 	*lenp = orig_size;
@@ -123,12 +122,11 @@ __wt_munmap(WT_SESSION_IMPL *session,
 	WT_UNUSED(mappingcookie);
 
 	WT_RET(__wt_verbose(session, WT_VERB_FILEOPS,
-	    "%s: unmap %p: %" PRIuMAX " bytes", fh->name, map, (uintmax_t)len));
+	    "%s: unmap %p: %zu bytes", fh->name, map, len));
 
 	if (munmap(map, len) == 0)
 		return (0);
 
 	WT_RET_MSG(session, __wt_errno(),
-	    "%s unmap error: failed to unmap %" PRIuMAX " bytes",
-	    fh->name, (uintmax_t)len);
+	    "%s unmap error: failed to unmap %zu bytes", fh->name, len);
 }
